Simplify setuid iptables exec: clear env, keep helpers in client.go
Use an empty child environment instead of filtering LD_*, and drop the
separate exec.go so the iptables spawn path stays in one file.
