+++ runtime/docker/firewall/iptables/client.go
@@ -3,7 +3,10 @@ package iptables
 import (
 	"bytes"
 	"fmt"
+	"os"
+	"os/exec"
 	"strings"
+	"syscall"
 
 	"github.com/charmbracelet/log"
 	"github.com/google/shlex"
@@ -213,4 +216,40 @@ func (c *IpTablesClient) ruleExists(af string, rule *definitions.FirewallRule) b
 	}
 
 	return false
+}
+
+// newIptablesCmd builds an iptables/ip6tables command that is safe to run from a
+// setuid-root containerlab binary (ruid=user, euid=0).
+//
+// iptables ≥ 1.8.8 exits 111 when getuid() != geteuid() because it loads match/
+// target shared libraries and therefore refuses to run under a setuid parent.
+// We isolate the fix to the child process via Credential so the long-lived
+// containerlab process keeps its real UID (important for $HOME file ownership).
+// Env is cleared so a poisoned library path cannot follow into iptables.
+func newIptablesCmd(name string, args ...string) *exec.Cmd {
+	cmd := exec.Command(name, args...)
+	// nil Env inherits the parent; empty means env -i.
+	cmd.Env = []string{}
+
+	if attr := iptablesSysProcAttr(os.Getuid(), os.Geteuid(), os.Getegid()); attr != nil {
+		cmd.SysProcAttr = attr
+	}
+
+	return cmd
+}
+
+// iptablesSysProcAttr returns SysProcAttr that sets the child's real/effective
+// UIDs to root when the parent is running setuid-root. Nil when no change is needed.
+func iptablesSysProcAttr(ruid, euid, egid int) *syscall.SysProcAttr {
+	if euid != 0 || ruid == 0 {
+		return nil
+	}
+
+	return &syscall.SysProcAttr{
+		Credential: &syscall.Credential{
+			Uid:         0,
+			Gid:         uint32(egid),
+			NoSetGroups: true,
+		},
+	}
 }
